pusht/diffusion_policy/dtl/requirements.txt
funcy==1.17
ipdb==0.13.9
lenses==0.5.0
matplotlib
pandas==1.4.3
parsimonious==0.9.0
pptree==3.1
PyYAML==6.0
scikit_learn
scipy
seaborn==0.11.2
tensorboardX==2.5.1
torch_scatter
torchvision
tqdm
attrs
tensorboard
